refactor: Validator 로직 분리
기존 하나의 Validator로 코치 및 못 먹는 음식 입력에 대한 검증을 했지만, 클래스로 분리
