+++ src/main/java/menu/util/validator/CoachValidator.java
@@ -1,4 +1,4 @@
-package menu.util;
+package menu.util.validator;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Set;
 import menu.util.constants.ExceptionMessage;
 
-public class Validator {
+public class CoachValidator {
 
     private static final String SEPARATOR = ",";
 
@@ -16,21 +16,12 @@ public class Validator {
     private static final int MINIMUM_NUMBER = 2;
     private static final int MAXIMUM_NUMBER = 5;
 
-    private static final int NOT_EAT_FOOD_MINIMUM_NUMBER = 0;
-    private static final int NOT_EAT_FOOD_MAXIMUM_NUMBER = 2;
-
     public static void validateCoach(String coachesByRest) {
         List<String> coaches = Arrays.asList(coachesByRest.split(SEPARATOR,-1));
         coaches.stream()
                 .forEach(coachName -> validateCoachName(coachName));
         validateCoachNumber(coaches);
         validateDuplicateCoachName(coaches);
-    }
-
-    public static void validateNotEatFoods(String notEatFoods) {
-        List<String> notEatFoodNames = Arrays.asList(notEatFoods.split(SEPARATOR,-1));
-        validateNotEatFoodNumber(notEatFoodNames);
-        validateDuplicateNotEatFoods(notEatFoodNames);
     }
 
     private static void validateCoachName(String name) {
@@ -55,16 +46,4 @@ public class Validator {
         }
     }
 
-    private static void validateNotEatFoodNumber(List<String> notEatFoods) {
-        if (!(notEatFoods.size() >= NOT_EAT_FOOD_MINIMUM_NUMBER && notEatFoods.size() <= NOT_EAT_FOOD_MAXIMUM_NUMBER)) {
-            throw new IllegalArgumentException(ExceptionMessage.NOT_EAT_FOOD_NUMBER);
-        }
-    }
-
-    private static void validateDuplicateNotEatFoods(List<String> notEatFoods) {
-        Set<String> distinct = new HashSet<>(notEatFoods);
-        if (notEatFoods.size() != distinct.size()) {
-            throw new IllegalArgumentException(ExceptionMessage.NOT_EAT_FOOD_DUPLICATE);
-        }
-    }
 }
